+++ src/js/submit/submitEvent.mjs
@@ -1,5 +1,5 @@
 import checkout from './checkout.mjs';
-import updateStats from '../js/stats/updateStats.mjs';
+import updateStats from '../stats/updateStats.mjs';
 
 export default function submitEvent(pokemon) {
   const submit = document.getElementById('submitButton');
@@ -45,7 +45,7 @@ export default function submitEvent(pokemon) {
         correct = true;
       }
 
-      checkout(correct);
+      checkout(correct, correctTypes);
       updateStats(correct);
     }
   }
